Rename: CreateNewPlaylist 파일명 변경 (CreatePlaylist), NewPlaylist 컴포넌트 삭제
